Wrapping up auth implementation
