+++ src/components/auth/LoginBtn.jsx
@@ -1,11 +1,14 @@
 "use client";
 
+import { redirect, useRouter } from "next/navigation";
 import { GoogleLogin } from "./action";
 
 export default function LoginBtn() {
+    const router = useRouter();
     const handleLogin = async () => {
         try {
             const response = await GoogleLogin();
+
             console.log("Login successful:", response);
         } catch (error) {
             console.error("Login failed:", error);
